% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWeather_API.R
\name{getCurrentOpenWeather}
\alias{getCurrentOpenWeather}
\title{Get current weather from OpenWeather (https://openweathermap.org/).}
\usage{
getCurrentOpenWeather(lat, long, KEY)
}
\arguments{
\item{lat}{character latitude}

\item{long}{character longitude}

\item{KEY}{character API key. Users must register to obtain it. See \href{https://openweathermap.org/appid}{link}.}
}
\description{
The forecast are retrived from Open Weather One Call API \href{https://openweathermap.org/api/one-call-api}{link}.
}
\details{
\code{getCurrentOpenWeather} returns a dataframe with date, temperature (Celsius scale), relative humidity (RH)(\%),
 windSpeed (m/s), windDegree, pressure, clouds, and conditionsID and description \href{https://openweathermap.org/weather-conditions#Weather-Condition-Codes-2}{link}

 In not all locations OpenWeather provides rain data, in this case the dataframe contains NA for the precipitation column.

 The `date` variable is given in local timezone.
}
\examples{
\dontrun{
key.ow <- "OW-API-KEY"
getForecastOpenWeather(lat = "37.9929600", long = "-1.5366100", KEY = key.ow)
}

}
