% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherbit_API.R
\name{getForecastWeatherBit}
\alias{getForecastWeatherBit}
\title{Get 48 h forecast from WeatherBit (https://www.weatherbit.io/) .}
\usage{
getForecastWeatherBit(lat, long, KEY)
}
\arguments{
\item{lat}{character latitude}

\item{long}{character longitude}

\item{KEY}{character API key. Users must register to obtain it. See \href{weatherbit.io/account/create}{link}.}
}
\description{
The forecast are retrived from \href{https://www.weatherbit.io/api/weather-forecast-120-hour}{WeatherBit 48 hour forecast}.
}
\details{
\code{getCurrentOpenWeather} returns a dataframe with date, temperature (Celsius scale), relative humidity (RH)(\%),
 windSpeed (m/s), windDegree, pressure, clouds, and conditionsID and description \href{https://openweathermap.org/weather-conditions#Weather-Condition-Codes-2}{link}

 In not all locations OpenWeather provides rain data, in this case the dataframe contains NA for the precipitation column.
}
\examples{
\dontrun{
key.wb <- "WB-API-KEY"
getForecastWeatherBit(lat = "37.9929600", long = "-1.5366100", KEY = key.wb)
}

}
