\name{meteo-package}
\alias{meteo-package}
\alias{meteo}
\docType{package}
\title{
meteo
}
\description{
}
\details{
\tabular{ll}{
Package: \tab meteo\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-16\cr
License: \tab What license is it under?\cr
}

}
\author{
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{

}
