% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openWeather_API.R
\docType{package}
\name{meteoForecast}
\alias{meteoForecast}
\title{meteoForecast: A package for obtain weather forecasts from well-known sources.}
\description{
Each forecast provider has its own particularities when it comes to data availability, this package aims to standardize the data in a single format.
}
\details{
At the moment the following variables are retrieved: temperature (Celsius scale), relative humidity (RH)(\%), rainfall (mm) and wind (m/s). Each forecast provider updates its forecasts at a different rate, which creates differences for the start and end date of the forecasts.
}
