\name{meteoForecast-package}
\alias{meteoForecast-package}
\alias{meteoForecast}
\docType{package}
\title{
meteoForecast
}
\description{
  meteoForecast
}
\details{

}
\author{
  Oscar Perpiñán, with contributions from Marcelo Almeida
}
\references{
  \url{http://www.meteogalicia.es/web/modelos/threddsIndex.action}
  \url{https://openmeteoforecast.org/wiki/Main_Page}
}

\keyword{ package }
\seealso{
}
\examples{

}
