\name{variables}
\alias{varsMG}
\alias{varsOM}
\docType{data}
\title{
variables
}
\description{
variables
}
\usage{
data(varsMG)
data(varsOM)
}
\format{
  \code{Name}, \code{Description} and \code{Units} of the variables
  available in the MeteoGalicia (\code{varsMG}) and OpenMeteo
  (\code{varsOM}) services.

}

\source{
  \url{http://www.meteogalicia.es/web/modelos/threddsIndex.action}
  \url{https://openmeteoforecast.org/wiki/Main_Page}
}

\examples{
data(varsMG)
data(varsOM)
}

\keyword{datasets}
